# Optimise sem_likelihood by memoising matrix generation functions