--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -47,7 +47,8 @@ Imports:
     rootSolve,
     stats,
     tidyr,
-    tidyselect
+    tidyselect,
+    xxhashlite
 Depends:
     R (>= 3.5)
 Language: en-US