--- a/bdsm.Rproj
+++ b/bdsm.Rproj
@@ -1,5 +1,4 @@
 Version: 1.0
-ProjectId: 1d0c72d4-caf9-4357-be21-068f589eb4a9
 
 RestoreWorkspace: No
 SaveWorkspace: No