… NAMESPACE
… man/sem_U1_matrix.Rd
… man/sem_sigma11_matrix.Rd
… man/sem_sigma12_matrix.Rd
… vignettes/references.bib