# Rationale

if you run the model space tests through a profiler:

```R
profvis::profvis(devtools::test(filter='model_space'))
```

and look at the CPU stats, you can make a couple observations:

1. Most of the `stats::optim` evaluation time is spent in the `sem_likelihood` function.

<img width="506" alt="image" src="https://github.com/user-attachments/assets/59cff6b3-d426-4caf-8dd8-a08111109183" />

2. Most of the `sem_likelihood` evaluation time is spent in just a few matrix generation functions, most notably `residual_maker_matrix` and `sem_sigma_matrix`.

<img width="873" alt="image" src="https://github.com/user-attachments/assets/d7c283be-c34e-49f2-941f-d3620dbc7813" />

<img width="895" alt="image" src="https://github.com/user-attachments/assets/e396fcb0-7890-429c-a466-9928548f733b" />

Now, let's look at the calls to these functions and see what kind of arguments they're run with:

<img width="682" alt="image" src="https://github.com/user-attachments/assets/554c702f-d5f5-4392-aadf-12c1e6320231" />

<img width="682" alt="image" src="https://github.com/user-attachments/assets/7930d555-e4a5-4145-99a3-b950c7441413" />

The pattern is clear: these functions are repeatedly called with the same arguments over and over. This is a perfect scenario for memoization.

# Memoization

Usually memoization is done by keeping a map of hashed parameters as keys and function returns as values. Whenever the function is called, the map is checked first to see whether the same arguments were used before. If so, the cached return value is retrieved instead of calling the function again.

In our case, through trial and error, I found that the best results were achieved by saving only the last function call. In most cases this gives us 90-100% cache hits, and we save time on hashing.

<img width="682" alt="image" src="https://github.com/user-attachments/assets/9140c573-57dc-438c-983e-17df985bf92e" />

<img width="682" alt="image" src="https://github.com/user-attachments/assets/95ddf25e-7c84-44d7-8f55-55d6e770bf75" />

To maximize memoization gains, I split `sem_sigma_matrix` into two functions and moved some calculations from `sem_likelihood` into them.

# So what kind of speed gains can we expect?

Hard to say without checking on real data, but on the `optim_model_space_params` test we're getting a 23% performance improvement as measured by the `bench` package:

<img width="509" alt="image" src="https://github.com/user-attachments/assets/215f7ea0-fa76-41ca-94ca-4ba6e111e06b" />


```
  expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
  <bch:expr> <bch:tm> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm>
1 original      14.1s  14.4s    0.0694        NA     5.03    20  1451       4.8m
1 memoized      10.9s  11.1s    0.0899        NA     3.94    20   876      3.71m
```

# Memory considerations

Measurements on the `optim_model_space_params` test actually show a drop in memory usage, which I believe is possible but can't confidently guarantee will always be true. The profvis profiler stopped working on me due to some bug or R version mismatch, and I gave up on fixing that for now ¯\\\_(ツ)\_\/¯.

<img width="601" alt="image" src="https://github.com/user-attachments/assets/ec0f331e-8874-4eba-bc78-7c00df6598e9" />

```
		Elapsed_Time_sec Total_RAM_Used_MiB Peak_RAM_Used_MiB
original           15.805                  0          24018654
memoized           11.308                0.1          16329712
```

# Further optimisations

For now I've dropped the initial plan of a C++ implementation and looked for a different approach for the following reasons:

1. `stats::optim` is already implemented in C/C++ and I haven't managed to create a reimplementation that matches existing results.
2. `sem_likelihood` is mostly vectorized except for some less CPU-intensive functions (e.g. `sem_B_matrix`, `sem_psi_matrix`).

We could still try rewriting `sem_likelihood` + matrix generation functions in C++ and compare results, but that's unknown territory at this point.

Some further ideas:
1. Remove `generate_params_vector`, which I think creates an unnecessary temporary matrix.
2. Reimplement just the aforementioned non-vector functions in C/C++.
3. Throw `SEM_likelihood.R` and `SEM_likelihood.R` at ChatGPT/Claude for C++ implementation.

# Attachments

[profvis-before.html.zip](https://github.com/user-attachments/files/20822664/profvis-before.html.zip)