@mszczepanczyk 2025-10-02
> @mszczepanczyk will we be working on this one, or is this already completed with https://github.com/mateuszwyszynski/bdsm/pull/126 and https://github.com/mateuszwyszynski/bdsm/pull/120?

It'd to have to be rewritten in C++ but I think it's worth it. Speed gains on the test data are massive. It's your call.